api/pyproject.toml
[tool.poetry]
name = "threatplaybook-api"
version = "4.0.0"
description = ""
authors = ["Abhay Bhargav <abhay@we45.com>"]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.68.1"
pyArango = "^1.3.5"
python-dotenv = "^0.19.0"
loguru = "^0.5.3"
pydantic = {extras = ["email"], version = "^1.8.2"}
PyJWT = "^2.1.0"
bcrypt = "^3.2.0"
uvicorn = "^0.15.0"
jmespath = "^0.10.0"

[tool.poetry.dev-dependencies]
black = "^21.8b0"
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
